mesh_server/docling/requirements.txt
requests>=2.25.0
flask>=2.0.0
flask-cors>=3.0.0
werkzeug>=2.0.0
python-dotenv>=0.19.0
pydantic>=1.8.0
docling>=0.1.0
transformers>=4.0.0